refactored primer3 to use the config directly instead of having all those properties passed in as args

+++ internal/repp/primer3.go
@@ -15,15 +15,6 @@ import (
 
 // primer3 is a utility struct for executing primer3 to create primers on a fragment
 type primer3 struct {
-	// Frag that we're trying to create primers for
-	f *Frag
-
-	// the Frag before this one
-	last *Frag
-
-	// the Frag after this one
-	next *Frag
-
 	// the target sequence
 	seq string
 
@@ -34,26 +25,25 @@ type primer3 struct {
 	out *os.File
 
 	// path to primer3 executable
-	primer3Path string
+	primer3Exec string
 
 	// path to primer3 config folder (with trailing separator)
+	// this is the path to primer thermodynamic parameters
 	primer3ConfDir string
+
+	// configuaration
+	config *config.Config
 }
 
 // newPrimer3 creates a primer3 struct from a fragment
-func newPrimer3(last, this, next *Frag, seq string, conf *config.Config) primer3 {
-	in, _ := os.CreateTemp("", "primer3-in-*")
-	out, _ := os.CreateTemp("", "primer3-out-*")
-
+func newPrimer3(seq string, conf *config.Config) primer3 {
+	// try to get primer3 executable from env if set
+	primer3Exec := getExecutable("PRIMER3_HOME", "bin", "primer3_core")
 	return primer3{
-		f:              this,
-		last:           last,
-		next:           next,
 		seq:            strings.ToUpper(seq),
-		in:             in,
-		out:            out,
-		primer3Path:    "primer3_core",
+		primer3Exec:    primer3Exec,
 		primer3ConfDir: conf.GetPrimer3ConfigDir(),
+		config:         conf,
 	}
 }
 
@@ -64,62 +54,58 @@ func newPrimer3(last, this, next *Frag, seq string, conf *config.Config) primer3
 // existing homology to begin with (the two nodes should share ~50/50)
 //
 // returning the number of bp that have to be artifically added to the left and right primers
-func (p *primer3) input(minHomology, maxHomology, maxEmbedLength, minSeqLength, pcrBuffer,
-	minPrimerLength, maxPrimerLength, optPrimerLength int,
-	maxHairpinMeltTempInCelsius, primerMinTm, primerMaxTm float64,
-) (addLeft, addRight int, err error) {
+func (p *primer3) input(f, prev, next *Frag) (addLeft, addRight int, err error) {
+	in, inErr := os.CreateTemp("", "primer3-in-*")
+	out, outErr := os.CreateTemp("", "primer3-out-*")
+
+	if inErr != nil || outErr != nil {
+		return 0, 0, multierr.Append(inErr, outErr)
+	}
+	p.in = in
+	p.out = out
+
 	// adjust the Frag's start and end index in the event that there's too much homology
 	// with the neighboring fragment
-	p.shrink(p.last, p.f, p.next, maxHomology, minSeqLength) // could skip passing as a param, but this is a bit easier to test
+	p.shrink(prev, f, next)
 
 	// calc the bps to add on the left and right side of this Frag
-	addLeft = p.bpToAdd(p.last, p.f, minHomology)
-	addRight = p.bpToAdd(p.f, p.next, minHomology)
+	addLeft = p.bpToAdd(prev, f)
+	addRight = p.bpToAdd(f, next)
 
-	start := p.f.start
-	length := p.f.end - start + 1
-
-	// sizes to make the primers and target size (min, opt, and max)
-	primerMin := minPrimerLength // defaults to 18
-	primerOpt := optPrimerLength
-	primerMax := maxPrimerLength // defaults to 23
+	start := f.start
+	length := f.end - start + 1
 
 	// check whether we have wiggle room on the left or right hand sides to move the
 	// primers inward (let primer3 pick better primers)
 	//
 	// also adjust start and length in case there's TOO large an overhang and we need
 	// to trim it in one direction or the other
-	leftBuffer := p.buffer(p.last.distTo(p.f), minHomology, maxEmbedLength, pcrBuffer)
-	rightBuffer := p.buffer(p.f.distTo(p.next), minHomology, maxEmbedLength, pcrBuffer)
+	leftBuffer := p.buffer(prev.distTo(f))
+	rightBuffer := p.buffer(f.distTo(next))
 
-	if length-leftBuffer-rightBuffer < minSeqLength {
+	if length-leftBuffer-rightBuffer < p.config.PcrMinFragLength {
 		leftBuffer = 0
 		rightBuffer = 0
 	}
 
 	// create the settings map from all instructions
-	file, err := p.settings(
-		p.seq,
-		p.primer3ConfDir,
+	settings := p.settings(
+		f.ID,
 		start,
 		length,
-		primerMin,
-		primerOpt,
-		primerMax,
 		leftBuffer,
 		rightBuffer,
-		maxHairpinMeltTempInCelsius,
-		primerMinTm,
-		primerMaxTm,
 	)
-	if err != nil {
-		return 0, 0, err
+	// write the settings to a buffer
+	var fileBuffer bytes.Buffer
+	for key, val := range settings {
+		fmt.Fprintf(&fileBuffer, "%s=%s\n", key, val)
 	}
-
-	if _, err := p.in.Write(file); err != nil {
+	fileBuffer.WriteString("=") // required at file's end
+	// then write them to the file
+	if _, err = p.in.Write(fileBuffer.Bytes()); err != nil {
 		return 0, 0, fmt.Errorf("failed to write primer3 input file %v: ", err)
 	}
-
 	return
 }
 
@@ -129,17 +115,18 @@ func (p *primer3) input(minHomology, maxHomology, maxEmbedLength, minSeqLength, 
 // the Frag and keep the overlap beneath the upper limit.
 // Only the end if shrunk. Only shifting from right side.
 // Otherwise, two neighboring fragments will both shrink and there won't be an overlap
-func (p *primer3) shrink(last, f, next *Frag, maxHomology int, minLength int) *Frag {
+func (p *primer3) shrink(last, f, next *Frag) *Frag {
 	var shiftInLeft int
 	var shiftInRight int
 
-	if distRight := f.distTo(next); distRight < -maxHomology {
+	if distRight := f.distTo(next); distRight < -p.config.FragmentsMaxHomology {
 		// there's too much homology on the right side, we should move the Frag's end inward
-		shiftInRight = (-distRight) - maxHomology
+		shiftInRight = (-distRight) - p.config.FragmentsMaxHomology
 	}
 
 	// make sure the fragment doesn't become less than the minimum length
-	canShrink := (f.end-shiftInRight)-(f.start+shiftInLeft) > minLength && len(f.Seq)-shiftInRight > shiftInLeft
+	canShrink := (f.end-shiftInRight)-(f.start+shiftInLeft) > p.config.PcrMinFragLength &&
+		len(f.Seq)-shiftInRight > shiftInLeft
 	if canShrink {
 		f.start += shiftInLeft
 		f.end -= shiftInRight
@@ -151,7 +138,7 @@ func (p *primer3) shrink(last, f, next *Frag, maxHomology int, minLength int) *F
 
 // bpToAdd returns the number of bp to add the end of the left Frag to create a junction
 // with the right Frag
-func (p *primer3) bpToAdd(left, right *Frag, fragsMinHomology int) int {
+func (p *primer3) bpToAdd(left, right *Frag) int {
 	if !left.couldOverlapViaPCR(right) {
 		return 0 // we're going to synthesize there, don't add bp via PCR
 	}
@@ -169,7 +156,7 @@ func (p *primer3) bpToAdd(left, right *Frag, fragsMinHomology int) int {
 	// eg: 5 bp distance leads to 2.5bp + ~10bp additonal
 	// eg: -10bp distance leads to ~0 bp additional:
 	// 		other Frag is responsible for all of it
-	b := math.Ceil(float64(fragsMinHomology) / float64(2))
+	b := math.Ceil(float64(p.config.FragmentsMinHomology) / 2)
 
 	return bpDist + int(b)
 }
@@ -179,17 +166,17 @@ func (p *primer3) bpToAdd(left, right *Frag, fragsMinHomology int) int {
 //
 // dist is positive if there's a gap between the start/end of a fragment and the start/end of
 // the other and negative if they overlap
-func (p *primer3) buffer(dist, minHomology, maxEmbedLength, pcrBuffer int) (buffer int) {
-	if dist > maxEmbedLength {
+func (p *primer3) buffer(dist int) (buffer int) {
+	if dist > p.config.PcrPrimerMaxEmbedLength {
 		// we'll synthesize because the gap is so large, add 100bp of buffer
-		return pcrBuffer
+		return p.config.PcrBufferLength
 	}
 
-	if dist < -minHomology {
+	if dist < -p.config.FragmentsMinHomology {
 		// there's enough additonal overlap that we can move this FWD primer inwards
 		// but only enough to ensure that there's still minHomology bp overlap
 		// and only enough so we leave the neighbor space for primer optimization too
-		return (-dist - minHomology) / 2
+		return (-dist - p.config.FragmentsMinHomology) / 2
 	}
 
 	return 0
@@ -199,40 +186,43 @@ func (p *primer3) buffer(dist, minHomology, maxEmbedLength, pcrBuffer int) (buff
 // can either use pick_cloning_primers mode, if the start and end primers' locations
 // are fixed, or pick_primer_list mode if we're letting the primers shift and allowing
 // primer3 to pick the best ones. One side may be free to move and the other not
-func (p *primer3) settings(
-	seq, p3conf string,
-	start, length, primerMin, primerOpt, primerMax, leftBuffer, rightBuffer int,
-	maxHairpinMeltTempInCelsius, primerMinTm, primerMaxTm float64,
-) (file []byte, err error) {
+func (p *primer3) settings(seqID string, start, length, leftBuffer, rightBuffer int) map[string]string {
+	var strictPrimerSelection string
+	if p.config.PcrPrimerUseStrictConstraints {
+		strictPrimerSelection = "0"
+	} else {
+		strictPrimerSelection = "1"
+	}
+
 	// see primer3 manual or /vendor/primer3-2.4.0/settings_files/p3_th_settings.txt
 	settings := map[string]string{
-		"SEQUENCE_ID":                          p.f.ID,
-		"PRIMER_THERMODYNAMIC_PARAMETERS_PATH": p3conf,
+		"SEQUENCE_ID":                          seqID,
+		"PRIMER_THERMODYNAMIC_PARAMETERS_PATH": p.primer3ConfDir,
 		"PRIMER_NUM_RETURN":                    "1",
-		"PRIMER_PICK_ANYWAY":                   "1",
-		"SEQUENCE_TEMPLATE":                    seq + seq,               // TODO
-		"PRIMER_MIN_SIZE":                      strconv.Itoa(primerMin), // default 18
-		"PRIMER_OPT_SIZE":                      strconv.Itoa(primerOpt),
-		"PRIMER_MAX_SIZE":                      strconv.Itoa(primerMax),
+		"PRIMER_PICK_ANYWAY":                   strictPrimerSelection,
+		"SEQUENCE_TEMPLATE":                    p.seq + p.seq,                             // TODO
+		"PRIMER_MIN_SIZE":                      strconv.Itoa(p.config.PcrPrimerMinLength), // default 18
+		"PRIMER_OPT_SIZE":                      strconv.Itoa(p.config.PcrPrimerOptimumLength),
+		"PRIMER_MAX_SIZE":                      strconv.Itoa(p.config.PcrPrimerMaxLength),
 		"PRIMER_EXPLAIN_FLAG":                  "1",
-		"PRIMER_MIN_TM":                        fmt.Sprintf("%f", primerMinTm),                 // defaults to 57.0
-		"PRIMER_MAX_TM":                        fmt.Sprintf("%f", primerMaxTm),                 // defaults to 63.0
-		"PRIMER_MAX_HAIRPIN_TH":                fmt.Sprintf("%f", maxHairpinMeltTempInCelsius), // defaults to 47.0
-		"PRIMER_MAX_POLY_X":                    "7",                                            // defaults to 5
-		"PRIMER_PAIR_MAX_COMPL_ANY":            "13.0",                                         // defaults to 8.0
+		"PRIMER_MIN_TM":                        fmt.Sprintf("%f", p.config.PcrPrimerMinTm),          // defaults to 57.0
+		"PRIMER_MAX_TM":                        fmt.Sprintf("%f", p.config.PcrPrimerMaxTm),          // defaults to 63.0
+		"PRIMER_MAX_HAIRPIN_TH":                fmt.Sprintf("%f", p.config.FragmentsMaxHairpinMelt), // defaults to 47.0
+		"PRIMER_MAX_POLY_X":                    "7",                                                 // defaults to 5
+		"PRIMER_PAIR_MAX_COMPL_ANY":            "13.0",                                              // defaults to 8.0
 	}
 
 	// if there is room to optimize, we let primer3 pick the best primers available
 	// with a range on either side of the fragment's start
-	// http://primer3.sourceforge.net/primer3_manual.htm#SEQUENCE_PRIMER_PAIR_OK_REGION_LIST
+	// https://primer3.org/manual.html#SEQUENCE_PRIMER_PAIR_OK_REGION_LIST
 	if leftBuffer > 0 || rightBuffer > 0 {
 		// V-ls  v-le               v-rs   v-re
 		// ---------------------------------
-		leftEnd := start + leftBuffer + primerMax
-		rightStart := start + length - rightBuffer - primerMax
+		leftEnd := start + leftBuffer + p.config.PcrPrimerMaxLength
+		rightStart := start + length - rightBuffer - p.config.PcrPrimerMaxLength
 		excludeLength := rightStart - leftEnd
 
-		if excludeLength >= 0 && excludeLength > primerMax {
+		if excludeLength >= 0 && excludeLength > p.config.PcrPrimerMaxLength {
 			settings["PRIMER_TASK"] = "generic"
 			settings["PRIMER_PICK_LEFT_PRIMER"] = "1"
 			settings["PRIMER_PICK_INTERNAL_OLIGO"] = "0"
@@ -241,15 +231,15 @@ func (p *primer3) settings(
 			// ugly undoing of the above in case only one side has buffer
 			if leftBuffer == 0 {
 				settings["SEQUENCE_FORCE_LEFT_START"] = strconv.Itoa(start)
-				settings["SEQUENCE_PRIMER_PAIR_OK_REGION_LIST"] = fmt.Sprintf(",,%d,%d ;", rightStart, rightBuffer+primerMax)
+				settings["SEQUENCE_PRIMER_PAIR_OK_REGION_LIST"] = fmt.Sprintf(",,%d,%d ;", rightStart, rightBuffer+p.config.PcrPrimerMaxLength)
 			}
 			if rightBuffer == 0 {
 				settings["SEQUENCE_FORCE_RIGHT_START"] = strconv.Itoa(start + length)
-				settings["SEQUENCE_PRIMER_PAIR_OK_REGION_LIST"] = fmt.Sprintf("%d,%d,, ;", start, leftBuffer+primerMax)
+				settings["SEQUENCE_PRIMER_PAIR_OK_REGION_LIST"] = fmt.Sprintf("%d,%d,, ;", start, leftBuffer+p.config.PcrPrimerMaxLength)
 			}
 
 			if settings["SEQUENCE_PRIMER_PAIR_OK_REGION_LIST"] == "" {
-				settings["SEQUENCE_PRIMER_PAIR_OK_REGION_LIST"] = fmt.Sprintf("%d,%d,%d,%d ;", start, leftBuffer+primerMax, rightStart, rightBuffer+primerMax)
+				settings["SEQUENCE_PRIMER_PAIR_OK_REGION_LIST"] = fmt.Sprintf("%d,%d,%d,%d ;", start, leftBuffer+p.config.PcrPrimerMaxLength, rightStart, rightBuffer+p.config.PcrPrimerMaxLength)
 			}
 			settings["PRIMER_PRODUCT_SIZE_RANGE"] = fmt.Sprintf("%d-%d", excludeLength, length)
 		}
@@ -262,19 +252,13 @@ func (p *primer3) settings(
 		settings["PRIMER_PRODUCT_SIZE_RANGE"] = fmt.Sprintf("%d-%d", length, length)
 	}
 
-	var fileBuffer bytes.Buffer
-	for key, val := range settings {
-		fmt.Fprintf(&fileBuffer, "%s=%s\n", key, val)
-	}
-	fileBuffer.WriteString("=") // required at file's end
-
-	return fileBuffer.Bytes(), nil
+	return settings
 }
 
 // run the primer3 executable against the input file
 func (p *primer3) run() (err error) {
 	p3Cmd := exec.Command(
-		p.primer3Path,
+		p.primer3Exec,
 		p.in.Name(),
 		"-output", p.out.Name(),
 		"-strict_tags",
@@ -291,7 +275,7 @@ func (p *primer3) run() (err error) {
 // parse the output into primers. add to fragment
 //
 // target is the target sequence we're building for. We need it to modulo the primer ranges
-func (p *primer3) parse(target string) (err error) {
+func (p *primer3) parse(target string) (primers []Primer, err error) {
 	fileBytes, err := os.ReadFile(p.out.Name())
 	if err != nil {
 		return
@@ -308,15 +292,18 @@ func (p *primer3) parse(target string) (err error) {
 	}
 
 	if p3Warnings := results["PRIMER_WARNING"]; p3Warnings != "" {
-		return fmt.Errorf("warnings executing primer3: %s", p3Warnings)
+		err = fmt.Errorf("warnings executing primer3: %s", p3Warnings)
+		return
 	}
 
 	if p3Error := results["PRIMER_ERROR"]; p3Error != "" {
-		return fmt.Errorf("failed to execute primer3 against %s: %s", file, p3Error)
+		err = fmt.Errorf("failed to execute primer3 against %s: %s", file, p3Error)
+		return
 	}
 
 	if numPairs := results["PRIMER_PAIR_NUM_RETURNED"]; numPairs == "0" {
-		return fmt.Errorf("failed to create primers using: \n%s", file)
+		err = fmt.Errorf("failed to create primers using: \n%s", file)
+		return
 	}
 
 	// read in a single primer from the output string file
@@ -357,19 +344,21 @@ func (p *primer3) parse(target string) (err error) {
 			Notes: notes,
 		}
 	}
-
-	p.f.Primers = []Primer{
+	primers = []Primer{
 		parsePrimer("LEFT", 0),
 		parsePrimer("RIGHT", 0),
 	}
-
 	return
 }
 
 func (p *primer3) close() (err error) {
 	// remove temporary input and output
-	err = multierr.Append(err, os.Remove(p.in.Name()))
-	err = multierr.Append(err, os.Remove(p.out.Name()))
+	if p.in != nil {
+		err = multierr.Append(err, os.Remove(p.in.Name()))
+	}
+	if p.out != nil {
+		err = multierr.Append(err, os.Remove(p.out.Name()))
+	}
 	return
 }
 
@@ -390,7 +379,7 @@ func hairpin(seq string, conf *config.Config) (melt float64) {
 
 	// see nnthal (no parameters) help. within primer3 distribution
 	ntthalCmd := exec.Command(
-		"ntthal",
+		getExecutable("PRIMER3_HOME", "bin", "ntthal"),
 		"-a", "HAIRPIN",
 		"-r",       // temperature only
 		"-t", "50", // gibson assembly is at 50 degrees
